+++ test/mock/mock_vidsrc.c
@@ -14,7 +14,7 @@ struct vidsrc_st {
 
 	struct vidframe *frame;
 	struct tmr tmr;
-	int fps;
+	double fps;
 	vidsrc_frame_h *frameh;
 	void *arg;
 };
@@ -71,7 +71,7 @@ static int mock_vidsrc_alloc(struct vidsrc_st **stp, const struct vidsrc *vs,
 
 	tmr_start(&st->tmr, 0, tmr_handler, st);
 
-	info("mock_vidsrc: new instance with size %u x %u (%d fps)\n",
+	info("mock_vidsrc: new instance with size %u x %u (%.2f fps)\n",
 	     size->w, size->h, prm->fps);
 
  out:
